Student performance in California schools: the results of the county 29

This database is a sample of the API program https://www.cde.ca.gov/re/pr/api.asp that ended in 2018. The sample is extracted from the data api of the package survey, related to the results of the county 29 (Nevada). The database contains information for the 418 schools of this county having at least 100 students. Missing information has been randomly (and voluntary) added to the awards and ell variables (4% and 7% respectively). Several variables have been voluntary categorized from their initial types.

api29

Format

A data.frame with 418 schools (rows) and 12 variables

cds

the school identifier

apicl_2000

the API score in 2000 classed in 3 ordered levels: [200-600],(600-800],(800-1000]

stype

the school type in a 3 ordered levels factor: Elementary, Middle or High School

awards

the school eligible for awards program ? Two possible answers: No or Yes. This variable counts 4% of missing information.

acs.core

the number of core academic courses in the school

api.stu

the number of students tested in the school

acs.k3.20

the average class size years K-3 in the school. This variable is stored in a 3-levels factor: Unknown, <=20, >20.

grad.sch

the percentage of parents with postgraduate education stored in a 3 ordered levels factor of percents: 0, 1-10, >10

ell

the percentage of English language learners stored in a 4 ordered levels factor: [0-10],(10-30],(30-50],(50-100]. This variable counts 7% of missing information.

mobility

the percentage of students for whom this is the first year at the school, stored in 2 levels: [0-20] and (20-100]

meals

the percentage of students eligible for subsidized meals stored in a 4 balanced levels factor (By quartiles): [0-25], (25-50], (50-75], (75-100]

full

the percentage of fully qualified teachers stored in a 2-levels factor: 1: For strictly less than 90%, 2 otherwise

Source

This database is a sample of the data api from the package survey.