OT_joint()

The function OT_joint integrates two models called (JOINT) and (R-JOINT) dedicated to the solving of recoding problems in data integration using optimal transportation of the joint distribution of outcomes and covariates.

OT_joint(
  datab,
  index_DB_Y_Z = 1:3,
  nominal = NULL,
  ordinal = NULL,
  logic = NULL,
  convert.num = NULL,
  convert.clss = NULL,
  dist.choice = "E",
  percent.knn = 1,
  maxrelax = 0,
  lambda.reg = 0,
  prox.dist = 0.8,
  prox.X = 0.8,
  which.DB = "BOTH"
)

Arguments

datab
A data.frame with at least four columns sorted in a non-specific order. One column must be a column of identifier for the two databases, where the names of the two databases must be ranked in ascending order (By example: 1 for the top database and 2 for the database from below, or more logically here A and B ...But NOT B and A!). One column (\(Y\) here but other names are allowed) must correspond to the target variable related to the information of interest to merge with its specific encoding in the database A (corresponding encoding should be so missing in the database B). In the same way, one column (\(Z\) here) corresponds to the second target variable with its specific encoding in the database B (corresponding encoding should be so missing in the database A). Finally, the input database must have at least one shared covariate with same encoding in A and B. Please notice that, if your data.frame has only four columns, that is to say with only one covariate, if this latter has NA, and unless user has previously imputed the corresponding missing information, the OT algorithm will only run complete cases.

index_DB_Y_Z
A vector of three column indexes. The first index must correspond to the index of the databases identifier column (DB identifier). The second index must correspond to the index of the target variable in the first database (A) while the third index corresponds to the index of column related to the target variable in the second database (B).

nominal
A vector of indexes that corresponds to the column indexes of all the nominal (not ordered) variables (DB identification and target variables included if it is the case for them).

ordinal
A vector of indexes that corresponds to the column indexes of all the ordinal variables (DB identification and target variables included if it is the case for them).

logic
A vector of indexes that corresponds to the column indexes of all the boolean variables of the data.frame.

convert.num
Indexes of the continuous (quantitative) variables. They will be automatically converted in ordered factors. By default, no continuous variables is assumed in the database.

convert.clss
A vector indicating for each continuous variable to convert, the corresponding desired number of levels. If the length of the argument convert_num exceeds 1 while the length of convert_clss equals 1 (only one integer), each discretization will count the same number of levels.

dist.choice
A character (with quotes) corresponding to the distance function chosen between: the euclidean distance ("E", by default), the Manhattan distance ("M"), the Gower distance ("G"), and the Hamming distance ("H") for binary covariates only.

percent.knn
Percent of closest neighbors taken in the computation of the cost matrix.

maxrelax
The maximum percentage of deviation from expected probability masses. It must be equal to 0 (default value) for the JOINT model, and equal to a strictly positive value for the R-JOINT model

lambda.reg
A coefficient measuring the importance of the regularization term. In the related reference (2), it corresponds to the R-JOINT model for a value other than 0 (default value))

prox.dist
A percentage (betwen 0 and 1) used to calculate the distance threshold below which an individual (a row) is considered as a neighbor of a given profile of covariates.

prox.X
A percentage (betwen 0 and 1) used to calculate the distance threshold below which two covariates' profiles are supposed as neighbors. If prox.X = 1, all profiles are considered as neighbors.

which.DB
A character indicating the database to complete ("BOTH" by default, for the prediction of \(Y\) and \(Z\) in the two databases), "A" only for the imputation of \(Z\) in A, "B" only for the imputation of \(Y\) in B.

Value

A list of 7 elements containing:

time_exe

Running time of the function

gamma_A

Estimation of \(\gamma\) for the completion of A. A cost matrix that corresponds to the joint distribution of \((Y^A,Z^A,X^A)\)

gamma_B

Estimation of \(\gamma\) for the completion of B. A cost matrix that corresponds to the joint distribution of \((Y^B,Z^B,X^B)\)

profile

A data.frame that gives all details about the remaining P profiles of covariates. These informations can be linked to the estimatorZA and the estimatorYB objects for a better interpretation of the results

res_prox

The outputs of the function proxim_dist

estimatorZA

An array that corresponds to estimates of the probability distribution of \(Z\) conditional to \(X\) and \(Y\) in database A. The number of rows of each table corresponds to the total number of profiles of covariates. The number of columns of each table corresponds to the number of levels of \(Y\). The row names of each table corresponds to the values of the covariates sorted by order of appearance in the merged database. The third element of the array is the possible level of \(Z\).

estimatorYB

An array that corresponds to estimates of the probability distribution of \(Y\) conditional to \(X\) and \(Z\) in database B. The number of rows of each table corresponds to the total number of profiles of covariates. The number of columns of each table corresponds to the number of levels of \(Z\). The row names of each table corresponds to the values of the covariates sorted by order of appearance in the merged database. The third element of the array is the possible level of \(Y\).

DATA1_OT

database A with imputed individual prediction on \(Z\) using OT

DATA2_OT

database B with imputed individual prediction on \(Y\) using OT

Details

A. THE RECODING PROBLEM IN DATA FUSION

Assuming that \(Y\) and \(Z\) are two variables that summarize a same latent information in two separate (no overlapping rows) databases A and B respectively, so that \(Y\) and \(Z\) are never jointly observed in A and B. Assuming also that A and B share a subset of common covariates X of any types (with same encodings in A and B) completed or not. Integrating these two databases often requires to solve the recoding problem observed between \(Y\) and \(Z\) by creating an unique database where the missing information of \(Y\) and/or \(Z\) is fully completed.

B. INFORMATIONS ABOUT THE ALGORITHM

As with the function OT_outcome, the function OT_joint provides a solution to the recoding problem by proposing an application of optimal transportation which aims is to search for a bijective mapping between the joint distributions of \((Y,X)\) and \((Z,X)\) in A and B (see (2) for more details). The principle of the algorithm is also based on the resolution of an optimization problem, which provides a \(\gamma\) solution (as called in (1) and (2)) which is an estimation of the joint distribution of \((X,Y,Z)\) according to the database to complete (see the argument which.DB for the choice of the database). While the models OUTCOME and R_OUTCOME integrated in the function OT_outcome require post-treatment steps to provide individual predictions, the algorithm JOINT direcly uses estimations of the conditional distributions \((Y|Z,X)\) in B and \((Z|Y,X)\) in A to predict the corresponding incomplete individuals informations of \(Y\) and/or \(Z\) respectively. This algorithm supposes that the conditional distribution \((Y|X)\) must be identical in A and B. Respectively, \((Z|X)\) is supposed identical in A and B. Estimations a posteriori of conditional probabilities \(P[Y|X,Z]\) and \(P[Z|X,Y]\) are available by profiles of covariates in output (See the objects estimatorYB and estimatorZA). Estimations of \(\gamma\) are also available according to the chosen transport distributions (See the arguments \(\gamma_A\) and \(\gamma_B\)).

The model R-JOINT gathers enrichments of the model JOINT and is also available via the function OT_joint It allows users to add a relaxation term in the algorithm to relax distributional assumptions (maxrelax>0), and (or) add also a positive regularization term (lamdba.reg>0) expressing that the transportation map does not vary to quickly with respect of covariates \(X\). Is is suggested to users to calibrate these two parameters a posteriori by studying the stability of the individual predictions in output.

C. TYPE OF REQUIRED INPUT DATABASE

The input database is a data.frame that must be saved in a specific form by users:

Two overlayed databases containing a common column of databases' identifiers (A and B, 1 or 2, by examples, encoded in numeric or factor form)

A column corresponding to the target variable with its specific encoding in A (For example a factor \(Y\) encoded in \(n_Y\) levels, ordered or not, with NAs in the corresponding rows of B)

A column corresponding to another target outcome summarizing the same latent information with its specific endoded in B (By example a factor \(Z\) in \(n_Z\) levels, with NAs in rows of A)

The order of the variables in the database have no importance but the indexes of the columns related to the 3rd columns previously described (ie ID, \(Y\) and \(Z\)) must be rigorously specified in the argument index_DB_Y_Z.

A set of shared common categorical covariates (at least one but more is recommended) complete or not (provided that the number of covariates exceeds 1) is required. On the contrary to the function OT_outcome, please notice, that the function OT_joint does not accept continuous covariates therefore these latters will have to be categorized beforehand or using the provided input process (see quanti).

The function merge_dbs is available in this package to assist user in the preparation of their databases, so please, do not hesitate to use it beforehand if necessary.

Remarks about the outcomes:

A target outcome can be categorical, in factor, ordered or not, discrete (with a finite number of values ONLY) but, notice that, if they are stored in numeric they will be automatically converted in ordered factors.

If a target outcome is incomplete, the corresponding rows will be automatically dropped during the execution of the function.

The type of each variables (including ID, \(Y\) and \(Z\)) of the database must be rigorously specified, in one of the four arguments quanti, nominal, ordinal and logic.

D. TRANSFORMATIONS OF CONTINUOUS COVARIATES

Continuous covariates with infinite numbers of values will have to be categorized beforehand their inclusions in the function. To assist users in this task, the function OT_joint integrates in its syntax a process dedicated to the categorization of continuous covariates. For this, it is necessary to rigorously fill in the arguments quanti and convert.clss. The first one informs about the indexes in database of the continuous variables to transform in ordered factor while the second one specifies the corresponding number of desired balanced levels (for unbalanced levels, users must do transformations by themselves). Therefore convert.num and convert.clss must be vectors of same length, but if the length of quanti exceeds 1, while the length of convert.clss is 1, then, by default, all the covariates to convert will have the same number of classes (transformation by quantiles), that corresponds to the value specified in the argument convert.clss. Please notice that only covariates can be transformed (not outcomes) and missing informations are not taken into account for the transformations. Moreover, all the indexes informed in the argument convert.num must also be informed in the argument quanti. Finally, it is suggested to declare all discrete covariates as ordinal factors using the argument ordinal.

E. INFORMATIONS ABOUT DISTANCE FUNCTIONS AND RELATED PARAMETERS

Each individual (or row) of a given database is here characterized by his covariates, so the distance between 2 individuals or groups of individuals depends on similarities between covariates according to the distance function chosen by user (via the argument dist.choice). Actually four distance functions are implemented in OT_joint to take into account the most frequently encountered situation (see (3)):

the Manhattan distance ("M")

the Euclidean distance ("E")

the Gower distance for mixed data (see (4): "G")

the Hamming distance for binary data ("H")

Let two profiles of covariates \(P_1\) (\(n_1\) individuals) and \(P_2\) (\(n_2\) individuals) will be considered as neighbors if \(dist(P_1,P_2) < prox.X \times max(dist(P_i,P_j))\) where \(prox.X\) must be fixed by user (\(i = 1,\dots,n_1\) and \(j = 1,\dots,n_2\)). This choice is used in the calculation of the JOINT and R_JOINT algorithms. In the same way, for a given profile of covariates \(Pj\), an individual i will be considered as a neighbor of \(Pj\) if \(dist(i,P_j) < prox.dist \times max(dist(i,P_j))\) where \(prox.dist\) will be fixed by user.

For more details about the related algorithms integrated in OT_joint, please consult (2).

References

Gares V, Dimeglio C, Guernec G, Fantin F, Lepage B, Korosok MR, savy N. (2019). On the use of optimal transportation theory to recode variables and application to database merging. The International Journal of Biostatistics. Volume 16, Issue 1, 20180106, eISSN 1557-4679 | https://doi.org/10.1515/ijb-2018-0106

Gares V, Omer J (2020) Regularized optimal transport of covariates and outcomes in data recoding. Journal of the American Statistical Association, DOI: 10.1080/01621459.2020.1775615

Anderberg, M.R. (1973), Cluster analysis for applications, 359 pp., Academic Press, New York, NY, USA.

Gower J.C. (1971). A general coefficient of similarity and some of its properties. Biometrics, 27, 623–637.

See also

Author

Gregory Guernec, Valerie Gares, Jeremy Omer

otrecod.pkg@gmail.com

Examples