indiv_grp_optimal()

This function assigns individual predictions related to recoding problems of data fusion by solving a linear optimization problem

indiv_grp_optimal(
  proxim,
  jointprobaA,
  jointprobaB,
  percent_closest = 1,
  which.DB = "BOTH"
)

Arguments

proxim
An object corresponding to the output of the function proxim_dist

jointprobaA
A matrix whose number of columns is equal to the number of modalities of the target variable \(Y\) in database A, and which number of rows is equal to the number of modalities of \(Z\) in database B. It gives an estimation of the joint probability \((Y,Z)\) in the database A. The sum of cells of this matrix must be equal to 1

jointprobaB
A matrix whose number of columns is equal to the number of modalities of the target variable Y in database A, and which number of rows is equal to the number of modalities of \(Z\) in database B. It gives an estimation of the joint probability \((Y,Z)\) in the database B. The sum of cells of this matrix must be equal to 1

percent_closest
A value between 0 and 1 (by default) corresponding to the fixed percent closest of individuals used in the computation of the average distances

which.DB
A character string (with quotes) that indicates which individual predictions compute: only the individual predictions of \(Y\) in B ("B"), only those of \(Z\) in A ("A") or the both ("BOTH" by default)

Value

A list of two vectors of numeric values:

YAtrans

A vector corresponding to the predicted values of \(Y\) in database B (numeric form) using the Optimal Transportation theory

ZBtrans

A vector corresponding to the predicted values of \(Z\) in database A (numeric form) using the Optimal Transportation theory

Details

A. THE RECODING PROBLEM IN DATA FUSION

Assuming that \(Y\) and \(Z\) are two variables that summarize a same latent information in two separate (no overlapping rows) databases A and B respectively, so that \(Y\) and \(Z\) are never jointly observed in A and B. Assuming also that A and B share a subset of common covariates X of any types (but same encodings in A and B) completed or not. Integrating these two databases often requires to solve the recoding problem observed between \(Y\) and \(Z\) by creating an unique database where the missing information of \(Y\) and \(Z\) is completed.

B. DESCRIPTION OF THE FUNCTION

The function indiv_grp_optimal is an intermediate function used in the implementation of a model called OUTCOME (and its enrichment R-OUTCOME, see the reference (2) for more details) dedicated to the solving of recoding problems in data fusion using Optimal Transportation theory. The model is implemented in the function OT_outcome which integrates the function indiv_grp_optimal in its syntax as a possible second step of the algorithm. The function indiv_grp_optimal can nevertheless be used separately provided that the argument proxim receives an output object of the function proxim_dist. This latter is available in the package and is so directly usable beforehand.

The function indiv_grp_optimal constitutes an alternative method to the nearest neighbor procedure implemented in the function indiv_grp_closest. As for the function indiv_grp_closest, assuming that the objective consists in the prediction of \(Z\) in the database A, the first step of the algorithm related to OUTCOME provides an estimate of \(\gamma\) which can is an estimation of the joint distribution \((Y,Z)\) in A. Rather than using a nearest neighbor approach to provide individual predictions,the function indiv_grp_optimal solves an optimization problem using the simplex algorithm which searches for the individual predictions of \(Z\) that minimize the computed total distance satisfying the joint probability distribution estimated in the first part. More details about the theory related to the solving of this optimization problem is described in the section 5.3 of (2).

Obviously, this algorithm runs in the same way for the prediction of \(Y\) in the database B. The function indiv_grp_optimal integrates in its syntax the function avg_dist_closest and the related argument percent_closest is identical in the two functions. Thus, when computing average distances between an individual i and a subset of individuals assigned to a same level of \(Y\) or \(Z\) is required, user can decide if all individuals from the subset of interest can participate to the computation (percent_closest = 1) or only a fixed part p (<1) corresponding to the closest neighbors of i (in this case percent_closest = p).

The arguments jointprobaA and jointprobaB can be seen as estimations of \(\gamma\) (sum of cells must be equal to 1) that correspond to estimations of the joint distributions of \((Y,Z)\) in A and B respectively.

References

Gares V, Dimeglio C, Guernec G, Fantin F, Lepage B, Korosok MR, savy N (2019). On the use of optimal transportation theory to recode variables and application to database merging. The International Journal of Biostatistics. Volume 16, Issue 1, 20180106, eISSN 1557-4679 | https://doi.org/10.1515/ijb-2018-0106

Gares V, Omer J (2020) Regularized optimal transport of covariates and outcomes in data recoding. Journal of the American Statistical Association, DOI: 10.1080/01621459.2020.1775615

See also

Author

Gregory Guernec, Valerie Gares, Jeremy Omer

otrecod.pkg@gmail.com

Examples