merge_dbs()

Overlay of two databases with specific outcome variables and shared covariates

merge_dbs(
  DB1,
  DB2,
  row_ID1 = NULL,
  row_ID2 = NULL,
  NAME_Y,
  NAME_Z,
  order_levels_Y = levels(DB1[, NAME_Y]),
  order_levels_Z = levels(DB2[, NAME_Z]),
  ordinal_DB1 = NULL,
  ordinal_DB2 = NULL,
  impute = "NO",
  R_MICE = 5,
  NCP_FAMD = 3,
  seed_func = sample(1:1e+06, 1)
)

Arguments

DB1
A data.frame corresponding to the 1st database to merge (top database)

DB2
A data.frame corresponding to the 2nd database to merge (bottom database)

row_ID1
The column index of the row identifier of DB1 if DB1 has one (no identifier by default)

row_ID2
The column index of the row identifier of DB2 if DB2 has one (no identifier by default)

NAME_Y
Name of the outcome (with quotes) in its specific scale/encoding from the 1st database (DB1)

NAME_Z
Name of the outcome (with quotes) in its specific scale/encoding from the 2nd database (DB2)

order_levels_Y
A vector of classes (with quotes) sorted in ascending order that permits to reorder the levels of \(Y\) in the 1st database (DB1) if necessary, when \(Y\) is stored as an ordinal factor (scale).

order_levels_Z
A vector of classes (with quotes) sorted in ascending order that permits to reorder the levels of \(Z\) in the 2nd database (DB2) if necessary, when \(Z\) is stored as an ordinal factor (scale).

ordinal_DB1
Vector of index of columns corresponding to ordinal variables in the 1st database (no ordinal variable by default)

ordinal_DB2
Vector of index of columns corresponding to ordinal variables in the 2nd database (no ordinal variable by default)

impute
A character (with quotes) equals to "NO" when missing data on covariates are kept (By default), "CC" for Complete Case by keeping only covariates with no missing information , "MICE" for MICE multiple imputation approach, "FAMD" for single imputation approach using Factorial Analysis for Mixed Data

R_MICE
Number of multiple imputations required for the MICE approach (5 by default)

NCP_FAMD
Integer corresponding to the number of components used to predict missing values in FAMD imputation (3 by default)

seed_func
Integer used as argument by the set.seed() for offsetting the random number generator (Random integer by default, only useful with MICE)

Value

A list containing 10 elements (11 if the argument impute equals "MICE"):

DB_READY

The database matched from the two initial databases with common covariates and imputed or not according to the impute option

Y_LEVELS

Levels retained for the target variable in the DB1

Z_LEVELS

Levels retained for the target variable in the DB2

REMOVE1

Labels of deleted covariates because of their different types observed from DB1 to DB2

REMOVE2

Removed factor(s) because of their different levels observed from DB1 to DB2

REMAINING_VAR

Labels of the covariates remained for the data integration using OT algorithm

IMPUTE_TYPE

A character with quotes that specify the method eventually chosen to handle missing data in covariates

MICE_DETAILS

A list containing the details of the imputed datasets using MICE when this option is enabled. Databases imputed for DB1 and DB2 according to the number of mutliple imputation selected (Only if impute = "MICE")

DB1_ID

A data.frame corresponding to the row identifier of DB1 after merging

DB2_ID

A data.frame corresponding to the row identifier of DB2 after merging

SEED

An integer used as argument by the set.seed function for offsetting the random number generator (random selection by default)

Details

Assuming that DB1 and DB2 are two databases (two separate data.frames with no overlapping part) to merge vertically, the function merge_dbs performs this overlay by checking the compatibility of the variables between the databases. Each database must contain a target variable (whose label must be filled in the argument Y for DB1 and in the argument Z for DB2 respectively, so that the final database in output will contain a target Y whose values will be missing in DB2 and another target Z whose values will be missing in DB1), a subset of shared covariates (by example, the best predictors of \(Y\) in DB1, and \(Z\) in DB2), and another possible subset of variables specific to each database. Each database can have a row identifier whose corresponding index of column must be assigned in the arguments row_ID1 for DB1 and row_ID2 for DB2. Nevertheless, by default, DB1 and DB2 have no row identifiers and the merging remains unchanged the order of rows in the two databases provided that \(Y\) and \(Z\) have no missing values. A rule for the overlay is that the declared first database (in the argument DB1) will be placed above the second one (declared in the argument DB2).

The function merge_dbs proposes a helpful tool for any user who wants to prepare their databases to data fusion by solving matching problems using algorithms in the package Otrecod via the functions OT_outcome or OT_joint. This function notably detects heterogenity between variables from one database to another and can so be useful as a preliminary step of data fusion using Optimal Transportation theory, where the sets of best predictors, specific to each base would be then selected using the function select_pred.

A. The function merge_dbs handles incomplete information, by respecting the following rules:

If Y or Z have missing values in DB1 or DB2, corresponding rows are excluded from the database before merging. Moreover, the data fusion remains unchanged the order of the rows in the 2 databases, and in the case of incomplete outcomes, if A and B have row identifiers, the corresponding identifiers are removed and these latters are stored in the objects DB1_ID and DB2_ID of the output.

Before overlay, the function deals with incomplete covariates according to the argument impute. Users can decide to work with complete case only ("CC"), to keep ("NO") or impute incomplete information ("MICE","FAMD").

The function imput_cov, integrated in the syntax of merge_dbs deals with imputations. Two approaches are actually available: the multivariate imputation by chained equation approach (MICE, see (3) for more details about the approach or the corresponding package mice), and an imputation approach from the package missMDA that uses a dimensionality reduction method (here a factor analysis for mixed data called FAMD (4), to provide single imputations. If multiple imputation is required (impute = "MICE"), the default imputation methods are applied according to the type of the variables. The average of the plausible values will be retained for a continuous variable, while the most frequent candidate will be remained as a consensus value for a categorical variable or factor (ordinal or not).

B. The function merge_dbs studies the compatiblities between A and B, of each shared covariate, by respecting the following rules:

The formats of Y and Z must be suitable. Categorical (ordered or not) factors are allowed. Numeric and discrete outcomes with a finite number of values are allowed but will be automatically converted as ordered factors using the function transfo_target integrated in the function merge_dbs.

Shared variables with incompatible formats between the two databases will be removed from the merging and the related label will be saved in output (REMOVE1).

Shared factors with incompatible levels (or number of levels) will be removed from the merging and the related label will be saved in output (REMOVE2).

Covariates whose names are specific to each database will be also deleted from the merging.

Notice that if certain important predictors have been improperly excluded from the merging, users can make the necessary transformations on these variables, and re-run the function. As a finally step, the function checks that all values related to \(Y\) in B are missing and inversely for \(Z\) in A.

References

Gares V, Dimeglio C, Guernec G, Fantin F, Lepage B, Korosok MR, savy N. (2019). On the use of optimal transportation theory to recode variables and application to database merging. The International Journal of Biostatistics. 0, 20180106 (2019),https://doi.org/10.1515/ijb-2018-0106

Gares V, Omer J. (2020) Regularized optimal transport of covariates and outcomes in data recoding. Journal of the American Statistical Association, DOI: 10.1080/01621459.2020.1775615

van Buuren, S., Groothuis-Oudshoorn, K. (2011). mice: Multivariate Imputation by Chained Equations in R. Journal of Statistical Software, 45(3), 1–67. urlhttps://www.jstatsoft.org/v45/i03/

Josse J, Husson F. (2016). missMDA: A Package for Handling Missing Values in Multivariate Data Analysis. Journal of Statistical Software, 70(1), 1–31. DOI: 10.18637/jss.v070.i01

See also

Author

Gregory Guernec

otrecod.pkg@gmail.com

Examples

### Assuming two distinct databases from simu_data: data_A and data_B
### Some transformations will be made beforehand on variables to generate
### heterogeneities between the two bases.
data(simu_data)
data_A = simu_data[simu_data$DB == "A",c(2,4:8)]; head(data_A)

#>       Yb1 Gender Treatment Dosage Smoking      Age
#> 1 [40-60[ Female     Trt A  Dos 3     YES 65.44273
#> 2 [20-40]   Male      <NA>  Dos 2      NO 51.78596
#> 3 [40-60[ Female   Placebo  Dos 2     YES 49.10844
#> 4 [40-60[ Female     Trt B  Dos 4    <NA> 56.43524
#> 5 [40-60[ Female     Trt A  Dos 4     YES 44.77365
#> 6 [20-40] Female     Trt A  Dos 1    <NA>       NA

data_B = simu_data[simu_data$DB == "B",c(3,4:8)]; head(data_B)

#>     Yb2 Gender Treatment Dosage Smoking      Age
#> 301   5 Female   Placebo  Dos 2     YES 44.58233
#> 302   1 Female     Trt B  Dos 4    <NA> 65.23921
#> 303   2 Female   Placebo   <NA>      NO 51.64228
#> 304   2 Female     Trt A   <NA>      NO 50.15125
#> 305   1 Female     Trt B  Dos 4     YES 61.53242
#> 306   1 Female     Trt B  Dos 3    <NA> 60.54819

# For the example, a covariate is added (Weight) only in data_A
data_A$Weight = rnorm(300,70,5)

# Be careful: your target variables must be in factor (or ordered) in your 2 databases
# Because it is not the case for Yb2 in data_B, we convert it.
data_B$Yb2    = as.factor(data_B$Yb2)

# Moreover, we store the Dosage covariate in 3 classes in data_B (instead of 4 classes in data_B)
# to make the encoding of this covariate specific to each database.
data_B$Dosage = as.character(data_B$Dosage)
data_B$Dosage = as.factor(ifelse(data_B$Dosage %in% c("Dos 1","Dos 2"),"D1",
                          ifelse(data_B$Dosage == "Dos 3","D3","D4")))

# For more diversity, we put this covariate at the last column of the data_B
data_B        = data_B[,c(1:3,5,6,4)]

# Ex 1: The two databases are merged and incomplete covariates are imputed using MICE
soluc1  = merge_dbs(data_A,data_B,
                    NAME_Y = "Yb1",NAME_Z = "Yb2",
                    ordinal_DB1 = c(1,4), ordinal_DB2 = c(1,6),
                    impute = "MICE",R_MICE = 2, seed_func = 3011)

#> DBS MERGING in progress. Please wait ... 
#> Y 
#> Z 
#> DBS MERGING OK 
#> -----------------------
#> SUMMARY OF DBS MERGING: 
#> Nb of removed subjects because of NA on targets: 0 ( 0 %) 
#> Nb of covariates removed because of differences between the 2 bases: 1 
#> Nb of covariates remained: 4 
#> Imputation on incomplete covariates: MICE 

summary(soluc1$DB_READY)

#>        DB              Y          Z            Age           Gender   
#>  Min.   :1.000   [20-40]:130   1   :145   Min.   :34.37   Female:414  
#>  1st Qu.:1.000   [40-60[:127   2   : 51   1st Qu.:46.83   Male  :286  
#>  Median :2.000   [60-80]: 43   3   : 35   Median :50.12               
#>  Mean   :1.571   NA's   :400   4   : 43   Mean   :50.12               
#>  3rd Qu.:2.000                 5   :126   3rd Qu.:53.51               
#>  Max.   :2.000                 NA's:300   Max.   :65.44               
#>  Smoking     Treatment  
#>  NO :374   Placebo:358  
#>  YES:326   Trt A  :209  
#>            Trt B  :133  
#>                         
#>                         
#>                         

# Ex 2: The two databases are merged and missing values are remained
soluc2  = merge_dbs(data_A,data_B,
                    NAME_Y = "Yb1",NAME_Z = "Yb2",
                    ordinal_DB1 = c(1,4), ordinal_DB2 = c(1,6),
                    impute = "NO",seed_func = 3011)

#> DBS MERGING in progress. Please wait ... 
#> Y 
#> Z 
#> DBS MERGING OK 
#> -----------------------
#> SUMMARY OF DBS MERGING: 
#> Nb of removed subjects because of NA on targets: 0 ( 0 %) 
#> Nb of covariates removed because of differences between the 2 bases: 1 
#> Nb of covariates remained: 4 
#> Imputation on incomplete covariates: NO 

# Ex 3: The two databases are merged by only keeping the complete cases
soluc3  = merge_dbs(data_A,data_B,
                    NAME_Y = "Yb1",NAME_Z = "Yb2",
                    ordinal_DB1 = c(1,4), ordinal_DB2 = c(1,6),
                    impute = "CC",seed_func = 3011)

#> DBS MERGING in progress. Please wait ... 
#> Y 
#> Z 
#> DBS MERGING OK 
#> -----------------------
#> SUMMARY OF DBS MERGING: 
#> Nb of removed subjects because of NA on targets: 0 ( 0 %) 
#> Nb of covariates removed because of differences between the 2 bases: 1 
#> Nb of covariates remained: 4 
#> Imputation on incomplete covariates: CC 

# Ex 4: The two databases are merged and incomplete covariates are imputed using FAMD
soluc4  = merge_dbs(data_A,data_B,
                    NAME_Y = "Yb1",NAME_Z = "Yb2",
                    ordinal_DB1 = c(1,4), ordinal_DB2 = c(1,6),
                    impute = "FAMD",NCP_FAMD = 4,seed_func = 2096)

#> DBS MERGING in progress. Please wait ... 
#> Y 
#> Z 
#> DBS MERGING OK 
#> -----------------------
#> SUMMARY OF DBS MERGING: 
#> Nb of removed subjects because of NA on targets: 0 ( 0 %) 
#> Nb of covariates removed because of differences between the 2 bases: 1 
#> Nb of covariates remained: 4 
#> Imputation on incomplete covariates: FAMD 

# Conclusion:
# The data fusion is successful in each situation.
# The Dosage and Weight covariates have been normally excluded from the fusion.
# The covariates have been imputed when required.