A simulated dataset to test the functions of the OTrecod package

The first 300 rows belong to the database A, while the next 400 rows belong to the database B. Five covariates: Gender, Treatment, Dosage, Smoking and Age are common to both databases (same encodings). The variables Yb1 and Yb2 are the target variables that summarize a same information saved in two distinct encodings, that is why, Yb1 is missing in the database B and Yb2 is missing in the database A.

simu_data

Format

A data frame with 700 observations on the following 8 variables.

DB

database identification, a character with 2 possible classes: A or B

Yb1

the target variable encoded in 3 ordered classes in DB A, missing in database B, stored as factor

Yb2

the target variable encoded in 5 ordered classes in DB B, missing in database A, stored as integer

Gender

a complete 2 class covariate saved as factor

Treatment

a 3 class covariates stored as a character with 2% of missing information

Dosage

a 4-levels factor, with 5% of missing information

Smoking

a 2 classes covariate stored as a character: NO for non smoker, YES otherwise. 10% of missing information

Age

a double corresponding to the age of participants in years. The last covariate of the database counts 5% of missing information

Source

randomly generated

Details

The purpose of the functions contained in this package is to predict the missing information on Yb1 and Yb2 in database A and database B using the Optimal Transportation Theory.

Missing information has been simulated to some covariates following a simple MCAR process.