transfo_target()

This function prepares the encoding of the target variable before running an algorithm using Optimal Transportation theory.

transfo_target(z, levels_order = NULL)

Arguments

z
A factor variable (ordered or not). A variable of another type will be, by default, convert to a factor.

levels_order
A vector corresponding to the values of the levels of z. When the target is ordinal, the levels can be sorted by ascending order. By default, the initial order is remained.

Value

The list returned is:

NEW

An object of class factor of the same length as z

LEVELS_NEW

The levels (ordered or not) retained for z

See also

Author

Gregory Guernec

otrecod.pkg@gmail.com

Examples

y      = rnorm(100,30,10)
aa     = transfo_target(y)

#> Your target y was numeric ... By default, it has been converted in factor of integers 
#> 41 remaining levels 

newlev  = unique(as.integer(y))
bb      = transfo_target(y,levels_order = newlev)

#> Your target y was numeric ... By default, it has been converted in factor of integers 

newlev2 = newlev[-1]
cc      = transfo_target(y,levels_order = newlev2)

#> Your target y was numeric ... By default, it has been converted in factor of integers 
#> Inappropriate number or declared labels of levels 
#> The default labels have been kept 

outco   = c(rep("A",25),rep("B",50),rep("C",25))
dd      = transfo_target(outco,levels_order = c("B","C","A"))
ee      = transfo_target(outco,levels_order = c("E","C","A","F"))

#> Inappropriate number or declared labels of levels 
#> The default levels have been kept 

ff      = transfo_target(outco)

outco2  = c(rep("A",25),NA,rep("B",50),rep("C",25),NA,NA)
gg      = transfo_target(outco2)
hh      = transfo_target(outco2,levels_order = c("B","C","A"))