verif_OT()

This function proposes post-process verifications after data fusion by Optimal Transportation algorithms.

verif_OT(
  ot_out,
  group.clss = FALSE,
  ordinal = TRUE,
  stab.prob = FALSE,
  min.neigb = 1,
  R = 10,
  seed.stab = sample(1:1e+06, 1)
)

Arguments

ot_out
An output object of the function OT_outcome or OT_joint

group.clss
A boolean indicating if the results related to the proximity between outcomes by grouping levels are requested in output (FALSE by default).

ordinal
A boolean that indicates if \(Y\) and \(Z\) are ordinal (TRUE by default) or not. This argument is only useful in the context of groups of levels (group.clss=TRUE).

stab.prob
A boolean indicating if the results related to the stability of the algorithm are requested in output (FALSE by default).

min.neigb
A value indicating the minimal required number of neighbors to consider in the estimation of stability (1 by default).

R
A positive integer indicating the number of desired repetitions for the bernoulli simulations in the stability study

seed.stab
An integer used as argument by the seed for offsetting the random number generator (Random integer by default). Only useful if the stability study is required.

Value

A list of seven objects is returned:

seed

The list of used random number generator. The first one is fixed by user or randomly chosen

nb.profil

The number of profiles of covariates

conf.mat

The global confusion matrix between \(Y\) and \(Z\)

res.prox

A summary table related to the association measures between \(Y\) and \(Z\)

res.grp

A summary table related to the study of the proximity of \(Y\) and \(Z\) using group of levels

hell

Hellinger distances between observed and predicted distributions

eff.neig

A table which corresponds to a count of conditional probabilities according to the number of neighbors used in their computation (Only the first ten values)

res.stab

A summary table related to the stability of the algorithm

Details

In a context of data fusion, where information from a same target population is summarized via two specific variables \(Y\) and \(Z\) (two ordinal or nominal factors with different number of levels \(n_Y\) and \(n_Z\)), never jointly observed and respectively stored in two distinct databases A and B, Optimal Transportation (OT) algorithms (see the models OUTCOME, R_OUTCOME, JOINT, and R_JOINT of the reference (2) for more details) propose methods for the recoding of \(Y\) in B and/or \(Z\) in A. Outputs from the functions OT_outcome and OT_joint so provides the related predictions to \(Y\) in B and/or \(Z\) in A, and from these results, the function verif_OT provides a set of tools (optional or not, depending on the choices done by user in input) to estimate:

the association between \(Y\) and \(Z\) after recoding

the similarities between observed and predicted distributions

the stability of the predictions proposed by the algorithm

A. PAIRWISE ASSOCIATION BETWEEN \(Y\) AND \(Z\)

The first step uses standard criterions (Cramer's V, Spearman's rank correlation coefficient) to evaluate associations between two ordinal variables in both databases or only one. When the argument group.clss = TRUE, these informations can be completed by those provided by the function error_group (available in the package), which is directly integrate in the function verif_OT. Assuming that \(n_Y > n_Z\), and that one of the two scales of \(Y\) or \(Z\) is unknown, this function gives additional informations about the potential link between the levels of the unknown scale. The function proceeds to this result in two steps. Firsty, error_group groups combinations of modalities of \(Y\) to build all possible variables \(Y'\) verifying \(n_{Y'} = n_Z\). Secondly, the function studies the fluctuations in the association of \(Z\) with each new variable \(Y'\) by using adapted comparisons criterions (see the documentation of error_group for more details). if grouping successive classes of \(Y\) leads to an improvement in the initial association between \(Y\) and \(Z\) then it is possible to conclude in favor of an ordinal coding for \(Y\) (rather than nominal) but also to emphasize the consistency in the predictions proposed by the algorithm of fusion.

B. SIMILARITIES BETWEEN OBSERVED AND PREDICTED DISTRIBUTIONS

When the predictions of \(Y\) in B and/or \(Z\) in A are available in the datab argument, the similarities between the observed and predicted probabilistic distributions of \(Y\) and/or \(Z\) are quantified from the Hellinger distance (see (1)). This measure varies between 0 and 1: a value of 0 corresponds to a perfect similarity while a value close to 1 (the maximum) indicates a great dissimilarity. Using this distance, two distributions will be considered as close as soon as the observed measure will be less than 0.05.

C.STABILITY OF THE PREDICTIONS

These optional results are based on the following decision rule which defines the stability of an algorithm in A (or B) as its average ability to assign a same prediction of \(Z\) (or \(Y\)) to individuals that have a same given profile of covariates x and a same given level of \(Y\) (\(Z\)).

Assuming that the missing information of \(Z\) in base A was predicted from an OT algorithm (the reasoning will be identical with the prediction of \(Y\) in B, see (2) and (3) for more details), the function verif_OT uses the conditional probabilities stored in the object estimatorZA (see outputs of the functions OT_outcome and OT_joint) which contains the estimates of all the conditional probabilities of \(Z\) in A, given a profile of covariates x and given a level of \(Y = y\). Indeed, each individual (or row) from A, is associated with a conditional probability \(P(Z= z|Y= y, X= x)\).

With the function OT_joint, the individual predictions for subject i: \(\widehat{z}_i\), \(i=1,\ldots,n_A\) are given using the maximum a posteriori rule: $$\widehat{z}_i= \mbox{argmax}_{z\in \mathcal{Z}} P(Z= z| Y= y_i, X= x_i)$$ While the function OT_outcome directly gives the individual predictions and the probablities \(P(Z= z|Y= y, X= x)\) are computed in a second step (see (3)).

For each subject i in database A, a new variable \(z_i'\) is simulate such that: $$z_i'\sim \mathcal{B}(P(Z= \widehat{z}_i|Y= y_i, X= x_i))$$ The average stability criterium is so calculated as: $$\mbox{Stab}_A = \frac{1}{n_A}\sum_{i=1}^{n_A} z_i'$$ This criterion can be repeated with R samples and the related mean and variance are given in output.

Nevertheless, it is possible that some of these conditional probabilities could be computed from only few individuals (because they are computed from a certain number of neighbor individuals for each covariates profile \(x\)) and lead to unreliable estimations. To avoid this problem, conditional probabilities can be removed from the stability criterion since they have been assessed from an insufficient number of subjects. In this way, the minimal number of subjects required for a conditional probability to participate to the stability estimation can be fixed a priori by filling in the argument min.neigb.

Notice that, these results are optional and available only if the argument stab.prob = TRUE. Finally, when the predictions of \(Z\) in A and \(Y\) in B are available, the function verif_OT provides in output, global results and results by database.

References

Liese F, Miescke K-J. (2008). Statistical Decision Theory: Estimation, Testing, and Selection. Springer

Gares V, Dimeglio C, Guernec G, Fantin F, Lepage B, Korosok MR, savy N. (2019). On the use of optimal transportation theory to recode variables and application to database merging. The International Journal of Biostatistics. Volume 16, Issue 1, 20180106, eISSN 1557-4679 | https://doi.org/10.1515/ijb-2018-0106

Gares V, Omer J (2020) Regularized optimal transport of covariates and outcomes in data recoding. Journal of the American Statistical Association, DOI: 10.1080/01621459.2020.1775615

See also

Author

Gregory Guernec

otrecod.pkg@gmail.com

Examples

### Example 1
#-----
# - Using the data simu_data
# - Studying the proximity between Y and Z using standard criterions
# - When Y and Z are predicted in B and A respectively
# - Using an outcome model (individual assignment with knn)
#-----
data(simu_data)
try1 = OT_outcome(simu_data, quanti = c(3,8), nominal = c(1,4:5,7), ordinal = c(2,6),
                 dist.choice = "G",percent.knn = 0.90, maxrelax = 0,
                 convert.num = 8, convert.clss = 3,
                 indiv.method = "sequential",which.DB = "BOTH",prox.dist = 0.30)

#> --------------------------------------- 
#> OT PROCEDURE in progress ... 
#> --------------------------------------- 
#> Type                     =  OUTCOME 
#> Distance                 =  Gower 
#> Percent closest knn      =  90 % 
#> Relaxation parameter     =  NO 
#> Relaxation value         =  0 
#> Individual pred process  =  Sequential 
#> DB imputed               =  BOTH 
#> --------------------------------------- 

ver1 = verif_OT(try1); ver1

#> $seed
#> [1] 875204
#> 
#> $nb.profil
#> [1] 232
#> 
#> $conf.mat
#>          predZ
#> predY       1   2   3   4   5 Sum
#>   [20-40] 254  49   1   1   1 306
#>   [40-60[   1  40  60  74 120 295
#>   [60-80]   1   1   1   1  95  99
#>   Sum     256  90  62  76 216 700
#> 
#> $res.prox
#>          N V_cram rank_cor
#> Global 700   0.73    0.880
#> 1st DB 300   0.72    0.870
#> 2nd DB 400   0.74    0.887
#> 
#> $res.grp
#> NULL
#> 
#> $hell
#>                 YA_YB ZA_ZB
#> Hellinger dist. 0.005 0.012
#> 
#> $eff.neig
#> NULL
#> 
#> $res.stab
#> NULL
#> 

if (FALSE) {

### Example 2
#-----
# - Using the data simu_data
# - Studying the proximity between Y and Z using standard criterions and studying
#   associations by grouping levels of Z
# - When only Y is predicted in B
# - Using an outcome model (individual assignment with knn)
#-----

data(simu_data)
try2 = OT_outcome(simu_data, quanti = c(3,8), nominal = c(1,4:5,7), ordinal = c(2,6),
                 dist.choice = "G",percent.knn = 0.90, maxrelax = 0,
                 convert.num = 8, convert.clss = 3,
                 indiv.method = "sequential",which.DB = "B",prox.dist = 0.30)

ver2 = verif_OT(try2, group.clss = TRUE, ordinal = TRUE); ver2

### Example 3
#-----
# - Using the data simu_data
# - Studying the proximity between Y and Z using standard criterions and studying
#   associations by grouping levels of Z
# - Studying the stability of the conditional probabilities
# - When Y and Z are predicted in B and A respectively
# - Using an outcome model (individual assignment with knn)
#-----

ver3 = verif_OT(try2, group.clss = TRUE, ordinal = TRUE, stab.prob = TRUE, min.neigb = 5,
       seed.stab = 991891); ver3

}